requirements.txt
pydantic>=2.8.0
python-dotenv>=1.0.1
requests>=2.32.2
tqdm>=4.66.4
# Optional (Week 3+)
# chromadb>=0.5.3
# faiss-cpu>=1.8.0.post1
# fastapi>=0.110.0
# uvicorn>=0.29.0
openai>=1.40.0
groq>=0.9.0
# anthropic>=0.34.0
pytest>=8.2.0








